"New Puzzle" button and Local Storage started
Basic RNG implemented.

Created a "New Puzzle" button that regenerates the puzzle when clicked. Altered the layout for the logged-in look to fit the button better.

Started working on local storage in the  "generatePuzzle()" function to store the puzzleArray and intArray when going between pages (still isn't working perfectly, but is close).

TODO: Fix the local storage, get the RNG to actually generate a puzzle (not just random numbers).